Implement proper serialization traits. Move files to proper packages. Add serialization.conf. Refactor for moved packages.
