+++ common-types/src/main/protobuf/com/improving/app/common/domain/memberId.proto
@@ -1,0 +1,9 @@
+syntax = "proto3";
+
+package com.improving.app.common.domain;
+import "scalapb/scalapb.proto";
+
+message MemberId {
+  option (scalapb.message).extends = "com.improving.app.common.serialize.PBMsgSerializable";
+  string id = 1;
+}
