+++ common-types/src/main/protobuf/com/improving/app/common/domain/tenantId.proto
@@ -1,0 +1,10 @@
+syntax = "proto3";
+
+package com.improving.app.common.domain;
+
+import "scalapb/scalapb.proto";
+
+message TenantId {
+  option (scalapb.message).extends = "com.improving.app.common.serialize.PBMsgSerializable";
+  string id = 1;
+}
