--- common-types/src/main/protobuf/com/improving/app/common/tenantId.proto
@@ -1,10 +1,0 @@
-syntax = "proto3";
-
-package com.improving.app.common;
-
-import "scalapb/scalapb.proto";
-
-message TenantId {
-  option (scalapb.message).extends = "com.improving.app.common.CborSerializable";
-  string id = 1;
-}
