+++ common-types/src/main/resources/serialization.conf
@@ -1,0 +1,5 @@
+akka.actor.serialization-bindings{
+  "com.improving.app.common.serialize.PBMsgSerializable" = proto
+  "com.improving.app.common.serialize.PBMsgOneOfSerializable" = proto
+  "com.improving.app.common.serialize.PBEnumSerializable" = proto
+}
