--- common-types/src/main/scala/com/improving/app/common/CborSerializable.scala
@@ -1,6 +1,0 @@
-package com.improving.app.common
-
-/**
- * The trait used to dictate what is to be serialized.
- */
-trait CborSerializable
