+++ common-types/src/main/scala/com/improving/app/common/serialize/PBEnumSerializable.scala
@@ -1,0 +1,3 @@
+package com.improving.app.common.serialize
+
+trait PBEnumSerializable extends scalapb.GeneratedEnum {}
