+++ common-types/src/main/scala/com/improving/app/common/serialize/PBMsgOneOfSerializable.scala
@@ -1,0 +1,3 @@
+package com.improving.app.common.serialize
+
+trait PBMsgOneOfSerializable extends scalapb.GeneratedSealedOneof {}
