+++ common-types/src/main/scala/com/improving/app/common/serialize/PBMsgSerializable.scala
@@ -1,0 +1,3 @@
+package com.improving.app.common.serialize
+
+trait PBMsgSerializable extends scalapb.GeneratedMessage {}
