+++ tenant/src/main/protobuf/com/improving/app/tenant/domain/improving-app-tenant-domain.proto
@@ -2,9 +2,9 @@ syntax = "proto3";
 
 package com.improving.app.tenant.domain;
 import "scalapb/scalapb.proto";
-import "com/improving/app/common/address.proto";
-import "com/improving/app/common/tenantId.proto";
-import "com/improving/app/common/memberId.proto";
+import "com/improving/app/common/domain/address.proto";
+import "com/improving/app/common/domain/tenantId.proto";
+import "com/improving/app/common/domain/memberId.proto";
 import "com/improving/app/tenant/domain/tenantTypeDefinitions.proto";
 
 option (scalapb.options) = {
@@ -13,44 +13,44 @@ option (scalapb.options) = {
 };
 
 message UpdateTenantName {
-  com.improving.app.common.TenantId tenant_id = 1;
+  com.improving.app.common.domain.TenantId tenant_id = 1;
   string new_name = 2;
-  com.improving.app.common.MemberId updating_user = 3;
+  com.improving.app.common.domain.MemberId updating_user = 3;
 }
 
 message TenantNameUpdated {
-  option (scalapb.message).extends = "com.improving.app.common.CborSerializable";
-  com.improving.app.common.TenantId tenant_id = 1;
+  option (scalapb.message).extends = "com.improving.app.common.serialize.PBMsgSerializable";
+  com.improving.app.common.domain.TenantId tenant_id = 1;
   string old_name = 2;
   string new_name = 3;
   MetaInfo meta_info = 4;
 }
 
 message UpdatePrimaryContact {
-  com.improving.app.common.TenantId tenant_id = 1;
-  com.improving.app.common.Contact new_contact = 2;
-  com.improving.app.common.MemberId updating_user = 3;
+  com.improving.app.common.domain.TenantId tenant_id = 1;
+  com.improving.app.common.domain.Contact new_contact = 2;
+  com.improving.app.common.domain.MemberId updating_user = 3;
 }
 
 message PrimaryContactUpdated {
-  option (scalapb.message).extends = "com.improving.app.common.CborSerializable";
-  com.improving.app.common.TenantId tenant_id = 1;
-  com.improving.app.common.Contact old_contact = 2;
-  com.improving.app.common.Contact new_contact = 3;
+  option (scalapb.message).extends = "com.improving.app.common.serialize.PBMsgSerializable";
+  com.improving.app.common.domain.TenantId tenant_id = 1;
+  com.improving.app.common.domain.Contact old_contact = 2;
+  com.improving.app.common.domain.Contact new_contact = 3;
   MetaInfo meta_info = 4;
 }
 
 message UpdateAddress {
-  com.improving.app.common.TenantId tenant_id = 1;
-  com.improving.app.common.Address new_address = 2;
-  com.improving.app.common.MemberId updating_user = 3;
+  com.improving.app.common.domain.TenantId tenant_id = 1;
+  com.improving.app.common.domain.Address new_address = 2;
+  com.improving.app.common.domain.MemberId updating_user = 3;
 }
 
 message AddressUpdated {
-  option (scalapb.message).extends = "com.improving.app.common.CborSerializable";
-  com.improving.app.common.TenantId tenant_id = 1;
-  com.improving.app.common.Address old_address = 2;
-  com.improving.app.common.Address new_address = 3;
+  option (scalapb.message).extends = "com.improving.app.common.serialize.PBMsgSerializable";
+  com.improving.app.common.domain.TenantId tenant_id = 1;
+  com.improving.app.common.domain.Address old_address = 2;
+  com.improving.app.common.domain.Address new_address = 3;
   MetaInfo meta_info = 4;
 }
 
@@ -72,31 +72,31 @@ message OrganizationsRemoved {
 }
 
 message ActivateTenant {
-  com.improving.app.common.TenantId tenant_id = 1;
-  com.improving.app.common.MemberId updating_user = 2;
+  com.improving.app.common.domain.TenantId tenant_id = 1;
+  com.improving.app.common.domain.MemberId updating_user = 2;
 }
 
 message TenantActivated {
-  option (scalapb.message).extends = "com.improving.app.common.CborSerializable";
-  com.improving.app.common.TenantId tenant_id = 1;
+  option (scalapb.message).extends = "com.improving.app.common.serialize.PBMsgSerializable";
+  com.improving.app.common.domain.TenantId tenant_id = 1;
   MetaInfo meta_info = 2;
 }
 
 message SuspendTenant {
-  com.improving.app.common.TenantId tenant_id = 1;
+  com.improving.app.common.domain.TenantId tenant_id = 1;
   string suspension_reason = 2;
-  com.improving.app.common.MemberId updating_user = 3;
+  com.improving.app.common.domain.MemberId updating_user = 3;
 }
 
 message TenantSuspended {
-  option (scalapb.message).extends = "com.improving.app.common.CborSerializable";
-  com.improving.app.common.TenantId tenant_id = 1;
+  option (scalapb.message).extends = "com.improving.app.common.serialize.PBMsgSerializable";
+  com.improving.app.common.domain.TenantId tenant_id = 1;
   MetaInfo meta_info = 2;
   string suspension_reason = 3;
 }
 
 message TenantRequest {
-  option (scalapb.message).extends = "com.improving.app.common.CborSerializable";
+  option (scalapb.message).extends = "com.improving.app.common.serialize.PBMsgSerializable";
   oneof sealed_value {
     UpdateTenantName update_tenant_name_value = 1;
     UpdatePrimaryContact update_primary_contact_value = 2;
@@ -107,7 +107,7 @@ message TenantRequest {
 }
 
 message TenantEvent {
-  option (scalapb.message).extends = "com.improving.app.common.CborSerializable";
+  option (scalapb.message).extends = "com.improving.app.common.serialize.PBMsgSerializable";
   oneof sealed_value {
     TenantNameUpdated tenant_name_updated_value = 1;
     PrimaryContactUpdated primary_contact_updated_value = 2;
